Add now pag algos.
