+++ pywhy_graphs/algorithms/__init__.py
@@ -1,2 +1,2 @@
-from .graph import *  # noqa: F403
+from .generic import *  # noqa: F403
 from .pag import *  # noqa: F403
