+++ pywhy_graphs/algorithms/tests/test_generic.py
